requirements.txt
py_clob_client_v2>=1.0.0
py-clob-client>=0.34.6
PyYAML>=6.0
requests>=2.32.0
